Add LayoutComponents as helpers for Sections, support SSR styling
This chnage firstly does some ground work to make sure that styled-components
works nicely with the SSR provided by next-js to make things easier to develop
with. For this, a customised _document.ts file needed to be made and some babel
config was created as well.

For other style related changes, a global stylesheet was created for basics like
font-family and components for organising the layout of a section were made more
customisable for their layouts.