Create non-functioning SectionPicker and do some component refactoring
This chanbge builds out the Header Component a bit more by refactoring
some of the existing components to allow them to be shared appropriately.
It also creates the SectionPicker Component which is just UI only at the
moment and doesn't do anything, but will have logic built out to it a bit
more later.