Add background blur and translucent background to the Header
This change adds a white background with a bit of opacity to the
Header Component, as well as setting a z-index on it so that it
renders properly with respect to the Avatar image on the landing.

The change also resolves some issues to do with the header animation
showing up early since it wasn't using the exact window size.