Adjust a spacing issue for the Header Avatar component
This change fixes up an issue where the Avatar component did not have
the right margin, and would have a bit of margin on the left of it
which would make it out of line with the rest of the content. This change
removes the extra margin to align it with the rest of the page.