Allow Section Components to have variable heights
The Landing component needs to be a bit larger than the rest of the
Section components so that the heading is centered, and normal Sections
will need to be less tall so that it doesn't lead to a huge amount ot scroll
between sections.

This change sets two sizes of Sections (big and normal) which is normal by
default but can be set explicitly for the Landing section.