Extract the logic for animating sections in and out to a Hook
Previously the logic that was used for animating in the About section
was pretty hard-coded and worked for the About section but wouldn't really
scale to the other sections since it used the overall window scroll distance
rather than looking at how much the page has been scrolled relative to the section.

This change consolidates the logic and then extracts it to it's own
hook which can be shared between sections without repeating code.