Refactor refs on Sections to let useScrollSpy hook track sections
This change installs the useScrollSpy hook to keep track of which
section is currently in the viewport. To allow this to work properly,
the hook needs a list of refs (one per section), so the refs on different
sections needed to be brought up a level to the index page and the passed down
to the Sections as well as the hook. With this, the Hook now knows which
Section is active and returns which index ref is active, which I just
pass down to the Section picker component to pick out the right
section from the list that is there.