Update animateOnScroll hook to remove exit animations
In a previous commit, I removed the exit animations but Sections would
still disappear out of the DOM when scrolled past. This was because
the hook was still measuring if the section was scrolled past and then
changing the state.

This change removes the bottom thresthold for sections and caches if a
section has already been scrolled in so that it doesnt animate out, only in.