Tweak useViewPortSize logic to better reflect how the usage would be
This chnage makes some changes to the useViewPortSize hook so that the
viewpost sizes are listed as enums, and can be picked from an array as
needed based on which is the most appropriate for the current viewport
size. This makes it much easier to say when a component should appear
at one breakpoint or between breakpoints.