Adjust useScrollDistanceHook to include scroll direction
Previously the useScrollDistance hook only returned the distance of
the viewport from the top of the window, which was exactly what it was supposed
to do.

In the SectionPicker, I wanted the SectionHeader to animate in different ways
depending on the ScrollDirection, which only involved swapping the variants
based on the scroll direction. To make this happen, this change adds as part of
the useScrollDistance hook, some logic to store what the direction of the
scroll is and return that to the component.