Call scrollHandler when useViewportSize hook mounts for immediate action
Previously the useViewportSizer hook would not fire on load since it would
only fire on resize, and so would need to be resized first before the viewport
size could be determined.

This change updates it so that on load the handler is fired to get an initial
reading on what the viewport is.