Create carousel component, configure offset styles
This change create the Carousel component that is the main component to
fill the Experience section. The Carousel needs to be a flowing component,
so it does not fit within the normal confines of a section with a container.
The approach in this change is to make it absolutely positioned and apply a ref-based
offset in pixels.