Adjust Carousel styles and add keys for items
This change makes some adjustments to the styles in the Carousel, mainly around
the Typography so that it is scaled appropriately for the boxes in the Carousel,
and refining the positioning logic in the Caruousel so that each time, which in
focus, is centered in the viewport. To accomplish this last bit, the resize observer
was added to the Carousel component so the offset can be adjusted everytime the
viewport changes, while also chaning the ref that stores the value to be state
so that it forces a repaint when updated.

Additionally some keys are added to the CarouselItems to prevent build issues.