Revise Section structure to remove absolute Experience positioning
Previously the Experience Section was absolutely positioned to allow
it to have an overflowing look relative to the container content. This
had its problems though since it messed up the layout for the rest of the
sections and made it difficult to position the icons.

This change makes the Container optional on the Section, so for the
Experience Section, it gets omitted and put on the Heading itself.
This make sure that the carousel can flow to the edge of the page
(minus some padding to align to the container size) without the need
for absolute positioning.