WIP: Create scrolling effect for Coverflow
This change adds a horizontally sticky element for the container
of the covers, so that the scroll event can still be detected as the
content stays still. With that in place, the Coverflow component tracks
how far the scrolling component has scrolled, maps that to which Cover
should be active, and then the Cover determines which one should be
the 'active' one acting as the center of the list.