Offset the Coverflow based on activeIndex so it stays centered
Previously the Coverflow content would not be centered when the active
index changes, so when the context scrolls the active item would keep getting
pushed futher right.

This change adds an offset based on how far the active index is from
the center so that the active element stays as centered as possible.