Major Refactor - Create StickySections and revert previous Sections
This change creates a new series of StickSection and related components
to act as a fundamental Component to allow Sections to be sticky and
animate in nicely. It pretty much replaces the old Section component
but doesnt remove it just yet.

Additionally this change reverts a lot of the work done in the Landing,
About, Experience, and Skills sections back to pretty much level one
again since I've tried to reimagine the main design and interaction.